WIP: Refactor projector creation to separate class ProjectorFactory
Move RenderImage and RenderStorage helper classes to separate files
