+++ src/main/java/bdv/viewer/render/ProjectorFactory.java
@@ -119,7 +119,7 @@ class ProjectorFactory
 			final ViewerState viewerState,
 			final RandomAccessibleInterval< ARGBType > screenImage,
 			final AffineTransform3D screenTransform,
-			final MultiResolutionRenderer.RenderStorage renderStorage )
+			final RenderStorage renderStorage )
 	{
 		/*
 		 * This shouldn't be necessary, with
@@ -144,11 +144,11 @@ class ProjectorFactory
 		else
 		{
 			final ArrayList< VolatileProjector > sourceProjectors = new ArrayList<>();
-			final ArrayList< MultiResolutionRenderer.RenderImage > sourceImages = new ArrayList<>();
+			final ArrayList< RenderImage > sourceImages = new ArrayList<>();
 			int j = 0;
 			for ( final SourceAndConverter< ? > source : visibleSources )
 			{
-				final MultiResolutionRenderer.RenderImage renderImage = renderStorage.getRenderImage( width, height, j );
+				final RenderImage renderImage = renderStorage.getRenderImage( width, height, j );
 				final byte[] maskArray = renderStorage.getMaskArray( j );
 				++j;
 				final VolatileProjector p = createSingleSourceProjector( viewerState, source, renderImage, screenTransform, maskArray );
